gelscreenmask in standalone file
Added gelscreenmask and updated the the unit and visual test to point to this.
Also added a test swf in the data folder for this and other tests to use.
This contents of gelscreenmask JS file should be moved into gelmodal once the changes in gelmodal
pass the gelscreenmask unit and visual tests.